remove sqlite3(ready to modularization), add bot and AI, add weapon module, add make list

+++ src/game/server/lastday/item/resource.h
@@ -23,7 +23,7 @@ enum ResourceList
 struct Resource
 {
     int64 GetResource(int ID);
-    void SetResource(int ID, int Num);
+    void SetResource(int ID, int64 Num);
     void ResetResource();
     int64 m_Metal;
     int64 m_Wood;
